Add '--no-fallback-messages' flag
Currently, fallback messages are always shown when the lemonade server is down.

I understand that it is quite important to tell what happens to the users.
However, I often found that the messages are bit annoying, especially when I
use lemonade as a clipboard program of neovim by manually assign `g:clipboard`.

So I added '--no-fallback-messages' flag to silence this warning like:

    $ echo 'hello' | pbcopy

    $ ./lemonade paste --host="127.0.0.1" --no-fallback-messages
    hello

    $ ./lemonade paste --host="127.0.0.1"
    2018/01/24 06:05:37 dial tcp 127.0.0.1:2489: getsockopt: connection refused
    2018/01/24 06:05:37 Fall back to localhost
    hello

This commit fixes #27